Add a node entry point
