+++ tsconfig.eslint.json
@@ -10,6 +10,7 @@
     "*.ts",
     "bin-src/**/*.*",
     "action-src/**/*.*",
+    "node-src/**/*.*",
     "src/**/*.*",
     ".storybook/*.*",
     "test-stories/**/*.*",
